pyproject.toml
[tool.poetry]
name = "flugel_challenge"
version = "0.1.0"
description = ""
authors = ["mcarello <mariano.carello@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.13"
djangorestframework = "3.13.1"
names = "0.3.0"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
